feat: display new label for cow amm
