+++ libs/ui/src/index.ts
@@ -33,6 +33,7 @@ export * from './pure/PercentDisplay'
 export * from './pure/CmsImage'
 export * from './pure/DismissableInlineBanner'
 export * from './pure/Input'
+export * from './pure/Badge'
 
 export * from './containers/CowSwapSafeAppLink'
 export * from './containers/InlineBanner'
